Separate regex expr into separate regex.py file
